Fixded Ticket #239 Fixed Bug: Sometimes link below cursor is not recognized
Cause : Anchor elements with javascript 'src' are not recognized as links while getting hit test result.
Solution: Now anchor elements with javascript 'src' will get recognized as links but WebHitTestResult.getExtra() will return blank as there is no usecase for the application to handle the javascript which there in the 'src' attribute.