Fix 2) from Video Features List.
2) Event Viewer:

a) When the application is loaded there is information about the version
that is the first thing that is seen when should be the FC. That
information can be available from the settings.

b) The colors of the text should be changed I think. Currently the
result is shown in white while the title in black. The tittle should be
dark gray and the result in black. We can ask Prisca about this.

c) Translate. I think that "Translating from" is not necessary, we
should leave "ENGLISH to CHINESE".

(references: #431)