Amazing!!!!. We have colored rings for single finger interaction.
Now the ring changes colors dynamically thanks to Velu's help and support. 

Tip 1: The color of the rings where set to dynamic with WebView.setCursorOuterColors(,,,) and WebView.setCursorInnerColors(,,,) and worked correctly, the problem is that was null the moment that was called in BrowserActivity onCreate. 

Quote from Velu's mail: "The only possibility I can think of is 'mNagiveClass' in WebView.java might be NULL at the time of calling setSearchHighlightColor(), setCursorOuterColors(), etc..."

Tip 2: setSearchHighlightColor() and setCursorOuterColors() are not refreshed unless the webview is invalidated WebView.invalidate();. This made the trick.