Variables/methods: ------------------
FC_SCROLL_SPEED variable on FC that sets the speed of the FC scrolling. Applied into dispatchTouchEventFC() 

scrollWebView() method calls showScrollCursor(quadrant, int direction, int distToX, int distToY) to check which cursor has to show on which case according to quadrant. If the distance to the corner is less that 200 the diagonal cursor is shown.

FIXME, diagonal cursor should be calculated with movement values rather than static distances. Currently that is calculated with isFCOutofBounds method.

getDeviceWidthHeight() and getFCLocation() on BrowserActivity provide the eyes of where is standing to scrollWebView().

TODO IMMEDIATE, disable media selection and execution timers. On next commit when I update them.
TODO, The scrolling cursors are not applied per page moving but per scrollTo. In fact is shown when the circle is moved outside the boundaries of the screen and not when really scrolling. It will be better to show them on page movement listener rather than page to be moved.

Assets:
-------
Scrolling cursors icons.