* FIXED BUG: Rings and Tips not properly located with or without stages (first, second, third).
Description: While dragging, when the mouse is on top of a link for example and the page was scrolled, althought the identifier was the same and the position also, the rect was in different positions when should send the same. As a result the rings and tabs are not located correctly. For example, this is the logcat for a same identified 4616948:

( 8778): first : Rect(93, 642 - 161, 671) identifier: 4616948 	//FIRST

( 8778): second : Rect(93, 488 - 161, 517) identifier: 4616948 	//SECOND
( 8778): second : Rect(93, 565 - 161, 594) identifier: 4616948

( 8778): third : Rect(93, 488 - 161, 517) identifier: 4616948 	//THIRD
( 8778): third : Rect(93, 565 - 161, 594) identifier: 4616948

I solved the problem by storing the identifier so only when is changed the rect calculation is done.